Feature/blueprint mcts (#71)
* Add blueprint based initial states. Add diversity sampling.

* Fixed namespace reinitialising bug

* Give reward for harvesting. Increase complexity exponent in calculating reward

* Implemented planning MCTS with the parallelisation infrastructure.

+++ src/datasetgen/mcts/conversation.py
@@ -50,7 +50,7 @@ class Message(BaseModel):
 class GenerationParameters(BaseModel):
     model: str
     n: int = 1
-    temperature: int = 0.7
+    temperature: float = 0.7
     max_tokens: int = 2048
     logit_bias: Optional[Dict[str, float]] = None
     stop_sequences: Optional[List] = None
